share/objects: Adjust armor on remaining monsters
Previously, armor aws adjusted during the game by using auras. This was
the case for both the difficulty setting and the level bonuses. From
this point on we will have the armor built in to the units, so to speak.
The armor values for Extreme are those that are adopted, while factoring
in the bonuses upon reaching various level milestones (i.e. Level 10,
20, and 30). This should enable scaling difficulty via life instead of
armor.