src/gem/spawn: Add difficulty scaling via life
This introduces difficulty scaling that is based upon life rather than
armor. The current values will be as follows:

- Easy: `55%`
- Normal: `70%`
- Hard: `85%`
- Extreme: `100%`

As one can see, Extreme is the primary target point, and will remain so
for balancing. The other difficulties remain, for the time being. It
should be noted that Easy and Normal will both feel harder in the early
game (i.e. before Level 30). However, the strategy used in Extreme will
be more directly applicable to the lower difficulties as armor remains
consistent.