src/gem: Units not attacking after selection/combination
Issues: #318

On rare occasions, a unit may not attack after a selection or
combination action. Previously, this behavior was exhibited by Topazes
before a changing them away from a bugged base unit. However, the same
problem is being encountered for combinations.

We go ahead and issue orders to `stop` for all units after they are
placed on the map, whether that be via a selection or a combination.