src/gem/slate: Fix Hold/Ancient taunt behavior
Issues: #279

Due to the removal of `PauseUnit ()` in 1.3.0, the behavior of Hold and
Ancient Slate taunt changed. Paused units would ignored various orders,
`attack` included. But after 1.3.0 units were no longer paused, and thus
were ordered to `attack` via the taunt.

It is not known whether Hold and Ancient slates were intentionally made
taunt immune, or if it was just dumb luck. However, this commit returns
the taunt behavior to that which existed prior to 1.3.0.