src: Add fix for stuck tooltip at map loading
Issues: #319

If a player's mouse is positioned just right (e.g. over a minimap
button) during initialization, the resulting tooltip can become
permanently stuck. It will not disappear, despite other tooltips
displaying properly underneath. This is a Warcraft 3, and not a map,
bug.

A solution to remove the tooltip after it has begun to misbehave has not
been found. However, this fix prevents the tooltip from misbehaving in
the first place.

This fix does present a few negative side effects:

- The function `ShowInterface ()` is utilized to hide the UI during map
  loaing and show it again on map start. This presents a possible conflict
  should an existing system have certain expectations regarding such
  behavior.
- On Ghost++ bots that make use of the `map_loadingame` setting, players
  will have their UI disabled until all other players have loaded. The
  primary negative here is the inability to use chat messages (e.g.
  Battle.net commands and map specific commands).

Due to the issue being rare, in conjunction with the side effects listed
above, Gem will not implement this fix as of yet. Once it is possible to
locally store settings, then the fix will be pushed out.