Revert "src/gem-3.1: Remove ownership change during placement/selection"
This reverts commit ae700ee4c7efef68491d9b67772a8f61798e8e87.

When using drag-select, units have a priority that is set in the Object
Editor (see 'Stats - Priority'). The lower the value, the higher the
priority. Units with higher priority are selected via drag-select over
units with a lower priority.

In Gem, this typically means that using drag-select over a group of
towers will give priority to slates. This is the desired behavior, given
that selecting an individual slate from a stack can be difficult at
times.

The problem arises when trying to select rocks during placement.
Previously, when the ownership of towers was changed, the only units
that could be selected via drag-select were the five being placed as
well as rocks. This made it easy to select rocks for removal. But if
towers do not change ownership, using drag-select to select a grouping
of a rocks becomes extremely difficult if a slate is present.

As such, this change is being reverted.