constants/interface: Fix version in upkeep area
Issues: #338

The removal of stretched widescreen in patch 1.29.0 meant that numerous
UI elements have less horizontal space.  This included the upkeep area.
Due to the version string (e.g. `1.4.2-alpha+fedfed1`) not being split,
some of it would not display as it would be too wide.

The solution is introducing a `\t` character before the dash.  This acts
as a hint for the game to split the version on whitespace that is not
rendered in client.