src: Add Run library
This is a fairly barebones library that hinges on `boolexpr` being run
via `TriggerAddCondition ()` and returning `true`.  In short, it will
display an error message if a thread has crashed or returns `false`.

A single `code` can be provided and will be run immediately.  A
`trigger` will be run when the attached event occurs, and will run all
added `code` in order.  A custom error message can be set that will be
displayed after the thread crashes or returns `false`.