src: Add Rule library
This library complements Run.  The purpose is to allow registration of a
trigger as a rule, then add events.  No conditions or actions are
intended to be added to this trigger.  Rather, when the event fire it
will utilize the Run library to call associated conditions in a manner
that allows detection of thread crashes and errors.