src: Add reworked Extra Chance
Issues: #242, #250, #268, #344

This reworks Extra Chance from the ground up.  All the old code is gone,
and what existed of the new has been refactored completely.

The following changes have been made:

- Upgrading Extra Chance has been removed.
- Extra Chance is active the entire round.
- Extra Chance for Perfects now applies to a whole type, increasing the
  chances for all qualities of a type.
- Extra Chance for Slates no longer gives the target slate.  Instead, it
  increases the chances for the components of said slate.
- A bonus that increases when a target is not hit and Extra Chance is
  used over consecutive rounds has been introduced.

All in all this both simplifies Extra Chance for the player, as well as
making cold (and hot streaks) less common in the long run.  Greats will
become more common again, in a manner that is much more acceptable for
balancing.  New downsides and upsides are introduced as well.