[Gem/Selection/Combine] Support 3x/4x
Issues: #128, #188

Previously, a player was provided with a single Combine button.  This
would always use the 'best' version of Combine, regardless of the
player's wishes.  If they wanted to double combine while having four of
a kind, too bad.

This commit introduces a new Combine (3x/4x) button that will exist
alongside the current Combine (2x).  The player will be able to choose
which style of combination they prefer.  At the moment both 3x and 4x
share the same icon.  That may change down the line if the two are to
ever be handled distinctly.