[Gem/Slate] Support slate stacking
Basically, we pull the functionality out of Gem 3.1.  In doing so we
also apply the updated stacking rules:

- A base slate cannot stack upon another slate of the same type or its
  combination.
- A combination slate cannot stack upon another slate of the same type
  or either of its component slates.

This effectivelly means that a base slate can now stack upon any other
base slate, except that of the same type.  For example, a Range can now
stack on a Damage but not another Range.  This should improve a player's
ability to cluster slates more to their liking, without the need to
reposition as often.