[Gem] Preload abilities
This preloads abilities that have commonly caused lag in-game when they
are first used.  The worry was that some of these would cause a marked
increase in load times; however, this worry was found to be without
merit.

Do note that there still may be further abilities that could benefit
from preloading.