[Gem/Extra Chance] Adjust message appearance
Closes: #354

This makes the 'Extra Chanced!' message appear immediately for Pefects,
which was the old behavior.  For slates, however, we will continue to
wait until the end of the placement round.

Additionally, we utilize line breaks to display the message for slates
over two lines, rather than using a second texttag.  This makes the
message immune to the effects of zooming out.