[Gem/Combination] Fix slate positioning issue
Closes: #361

The function `CreateUnit ()` attempts to realign newly created units on
`0.250` boundaries by rounding the provided coordinates.  In most cases,
this presents no issue as the majority of units in Gem will be already
aligned.  Slates are the exception.

It was possible, albeit rare, for a base slate positioned closely to
other slates in a non-stacking fashion to have its resultant combination
slate be realigned closer to existing slates and thus be considered as
stacking.  It would subsequently be moved.