[Gem/Extra Chance] Rework for Slates
We leverage the work done for Perfects to make adjusting the values for
Slates easier.  Additionally, whereas only the chances for a single
base slate were increased, we now increase the chances for both base
slates within a pair.  The target base slate is increased more than than
its cooresponding partner.

In a sense, one now chances for a combination slate with the target base
slate of their choice.  However, one will be offered more of the gems
that can make either base slate, instead of just the one.  The current
values maintain the rarity distiction betwen two-piece and three-piece
slates.