[Src/Board] Refactor title and board contents
Closes: #19
Closes: #92
Closes: #304
Closes: #353
Closes: #358
Closes: #362

The board is completely reworked.

The board title, which had previously been underutilized, is now much
more informative.  It lists a player's rank, level, lives, and extra
chance information.  That last bit of information is highly important
given the rework to Extra Chance in 1.5.0.

As for the board contents, a new column has been added: DPS.  This
tracks a player's DPS on a given round.  Additionally, the clock has
been removed.  Shifting it to the title was found to not be possible, as
seeing the title shift dimensions constantly was deemed too distracting.
The same consideration was made when not including DPS on the title.

Upon any player reaching the Damage Test a new column is added: Test.
This will display the damage dealt to the Damage Test, which became
difficult or impossible to see due to the shift away from stretched
widescreen.  Upon killing the Damage Test, the player's time will be
printed.