[Src/Board] Rework DPS text
Two needed changes to the display of the DPS values were identified.
First, decimals were simply too distracting, so we remove them entirely
and only display integers.  Second, DPS for players who have finished a
round but not yet started the next will be shown as grey.  This too will
make it easier to grok the board quickly.