[Gem] Detect cheats in single-player
Close: #374

This does not prevent people from using them in single-player.  That is
not the goal.  The goal is to allow someone cruising through the replay
the ability to determine that they have been used.  This will be
displayed upon finishing the Damage Test by having the time for the
player be shown in grey.

As far as the detection itself, this is done without creating visible
units to the player and in a fashion that should not alter gameplay at
all.  Functionality is quite spartan, and only a single function to
determine if cheats were detected is provided.  Upon detection, the
system disables itself.

The following cheats are detected:

- Invincibility: `whosyourdaddy`.
- Infinite mana: `thereisnospoon`.
- Gold: `keysersoze`.
- Lumber: `leafittome`.
- Gold and lumber: `greedisgood`.
- Fast building: `warpten`.