[Gem] Ensure games list player count is consistent
Close: #375

As of patch 1.30.2, the games list shows the current number of players
in a lobby.  However, it seems to rely upon information in both the
`war3map.w3i` and `config ()` function.  The values used seems to vary
at times.  As such, for Gem this is either displayed as `#/9` or `#/11`.
The first is most likely pulled from the `config ()`.  The second from
the `war3map.w3i`.

We go ahead and remove a few player declarations from the `war3map.w3i`,
thus aligning it with what is in present within `config ()`.  This seems
to address the issue.  Trying to target `#/8` encounters a similar
issue, where either `#/8` or `#/9` is used.