[Src/Gem] Introduce a basic game identifier
Close: #382

This is not a unique game identifier.  That is not its purpose.  The
point is to allow verification that a screenshot pairs to a replay in an
easy, albeit not foolproof, way.

As far as making use of this, the identifier is added to the board title
during replays and for players who have killed the Damage Test.