[Gem/Rank] Ensure accurate damage on Test when leaking
Close: #377

When a player who leaks the Damage Test subsequently leaves the game the
damage recorded jumps to its maximum life.  This happens because when a
unit is removed via `RemoteUnit ()` its life is zeroed.  The Rank module
picks up on this and was simply updating the damage done.

Now the correct behavior is determined, and damage is no longer updated
after a player has 'failed'.