[Gem 3.1] Improve creep attack detection further
Issue: #373

Rather than detecting when a creep begins an attack action, we detect
when a creep is given the attack order.  This has the distinct advantage
of preventing a player from using rocks as a barrier and preventing the
creeps from ever getting into attack range at all.

Additionally, we handle both the first and last checkpoints as well.
This resolves an issue where creeps would teleport to the middle of the
map.  In actual gameplay this would most likely never be encountered.