[Gem 3.1] Update range check image and accuracy
Issue: #386

We switch from the default selection circle to the basic area of effect
circle.  This exhibits superior edges at all sizes.

Of particular note is that the image used is `256x256`, whereas the
circle itself does not extend to the edges.  It is closer to `242x242`,
so an adjustment to the radius needs to be made.

All of the default selection circles need this adjustment; however, the
exact value may differ.