[Gem 3.1] Prevent slates from moving outside area
Issue: #190

The primary concern here is moving a slate into another player's area.
This has the potential to cause havoc, ranging from destroying their
structures to stealing their kills.

We leverage the same check used to constrain the builder to the player's
area, but apply it to slates that are being teleported.