[Commands/Zoom] Enhance zoom for replays
Issue: #400

This change does two things:

1. It disables the zoom command for replays.  This should make the
   viewing experience better in such situations.
2. It sets a default zoom of `2500`.  This is only for replays.  If the
   player wants the Warcraft III default zoom of `1650` they can press
   Insert or Delete.  This change should prevent the zoom from being
   stuck at some small value (e.g. `400`) from someone placing a slate.

Of particular note is that the same risks of using `GetLocalPlayer ()`
and desyncs apply to the use of game status as well.  The only
difference is that there may not be a desync (as in the case of
replays), but the behavior will essentially be buggy due to the game
state in the replay not matching that produced during the replay
simulation.