[Gem/Extra Chance] Properly charge/refund cost
The bug here is that switching Extra Chance targets would fail to
charge/refund the difference, if there was one.  When active, it would
switch targets without acknowleding cost at all.

As such, a player could switch from a cheaper target to a more expensive
one for free.  They could even use Cancel and be refunded extra gold.
It was also possible to be refunded less gold than inteded as well.