[Src/Cheats] Disable module
Issue: #409

This resolves an issue introduced in Gem TD+ 1.5.2 that cause online
(either Battle.net or LAN) games involving a single player to create
invalid replays.  Well, that is not entirely accurate, but it is a good
way to describe the issue to players.

In reality, the map itself was entering a different state than that
played in the game in the above situation.  The cheats detecton module
was enabling in the replay even though it was not activated in the game
(due to being online).  As such, the replay was accurate; but, it was
impossible to watch it in the game client.