[Src/Settings/Setup] Restore user control
Issue: #406

This involves an issue that cannot be reproduced.  However, it is
similar to that addressed in #159 that dealt with dialog windows and the
loss of user control.  This one, however, occurs at map start.

As a result, we hit it with every native remotely related to control and
user interface.  Maybe this resolves the issue.  Maybe not.  It cannot
hurt.