[Src/Main] Disable HCL
HCL does not need to be active anymore.  It was being left in mostly as
a means to innoculate the map against improper use of HCL.  However, the
map seems to be primarily played on the latest Warcraft III patches.
Hence, no need to be worried about bot specific behavior.