[Lib] Add Wurst Standard Library as submodule
[Gem] Add Wurst supportive files

The `_build` directory is needed for VS Code, as it is automatically
generated.  Gem provides links to the latest `common.j` and `blizzard.j`
stored within the repository.

As far as `wurst.dependencies`, this file is needed in the project's
root directory to satisfy the `build` command.  It is needed within the
`wurst` directory to satisfy Wurst's language server when using VS Code.