[Gem] Support Map 0.6.4
Primarily, this adds additional support for Wurst.  A change unrelated
to Wurst involves optimized versions of the map being built into a
separate directory.

The `_build` directory is needed for VS Code, as it is automatically
generated.  Gem provides links to the latest `common.j` and `blizzard.j`
stored within the repository.

As far as the `wurst` directory and `wurst.dependencies` file, Map
automatically handles their exitstence, so we just have Git ignore them.