[Gem/Extra Chance] Rework design
Issue: #380
Issue: #411

The general design of Extra Chance is simplified.  Previously, it was a
goal to only adjust values directly within the chance table.  This,
however, left little flexibility when approaching balance.

Instead, the new 'On_Pre_Placement' event is utilized.  This allows
adjusting the chance table directly before a gem is placed.  As such, it
is now possible to treat Extra Chance 'events' separately; when no such
event occurs, the default chance tables are used.  This allows
performing all rolls for Extra Chance at the start of the placement
round.  This opens up the possibility of same luck and fixed seeds.

In terms of behavior, things mostly remain the same for Perfects.  A
buff for the target Perfect was introduced.  As for Slates, the behavior
is different.  Instead of every roll being entirely independent, now
rolls are handled conditionally.  The first Extra Chande event will roll
a normal piece from either the target slate or its associated paired
slate.  The next event will roll a flawed gem that is part of the same
slate.  This process repeats for however many events exist.  This
ensures behavior that should frustrate players less, and make Extra
Chance for Slates more rewarding in general.

In regards to the specific issue of high bonuses not feeling rewarding,
the design for both Perfects and Slates has changed.  A bias has been
introduced that grows exponentially as bonus increases.  This bias skews
the numbers towards particular events (e.g. the target Perfect and
one-hit combination slates).  Additionally, the maximum bonus has been
increased to `10x`, where the player is essentially guaranteed nearly
all Perfects or a one-hit combination slate.