[Gem/Placement] Detect `Esc` key while building
Issue: #416

In order to cancel construction, one must select the building being
constructed.  This, however, is error prone.  There is a short time
window in which to do this.  One also must select the building.

An alternative option would be to automatically select the unit.
However, that prevents giving the building multiple orders.  As such, we
detect usage of the escape key.  If the builder is selected and
building, we cancel construction of the building.