[Objects/Builder] Change build style to Orc
Issue: #416

The old style utilize Undead construction.  In short, the unit warped in
out of nothing.  This allowed players to 'fast build'.  However, this
style of building is difficult to use when latency is high.  Removing it
makes games with high ping less problematic for players, and should help
the multiplayer situation.  The new style will utilize Orc construction.