[Gem 3.1] Shrink movement region sizes
Issue: #11

First, a bit of background.  A `region` is made up internally by cells.
Cells have dimensions of `32 x 32`, and exclude the entire top and right
edges.  Thus, one could specify the point `(0, 0)` and have the entire
cell marked.

The original style of specifying the movement regions used `rect` with
dimensions of `64 x 64`.  However, unlike cells, a `rect` is inclusive
on all edges.  In turn, this ends up creating regions `96 x 96` in size,
as unintended cells are included.  As a result, units were not traveling
as far into the checkpoint as desired when approaching from the north or
east.

We reduce the region sizes (to less than `64 x 64`) In this case, `0.1 x
0.1`; although, any size that will pull in the four cells around the
center point will do.  The result is a `64 x 64` region, as originally
intended.