[Gem/Placement] Fix/improve builder movement
Builder movement properly handles all gem placement scenarios.  In
short, those that were not aligned to `128` sized squares were ignored.
This obviously allowed the builder to move to locations that should not
have been considered.

On top of it, the order of a few checks are swapped to improve
performance.