[Gem 3.1/Movement] Add waypoints for air units
Issue: #11

This introduces temporary waypoints for air units.  This is achieved by
issuing periodic orders for said units to move a short distance towards
the next checkpoint.  This helps fight against the pathing engine's
tendency to have units drift from the straight line.

Note that this is only needed for air units.  It is intended for ground
units to have their pathing dictated by mazing.  Applying waypoints for
them is unnecessary.  And, also more difficult!