[Gem/Extra Chance] Decrease bonus instead of clearing
Issue: #380

For a bonus of `3x` or below, this changes nothing.  Hitting the target
will continue to clear the bonus.  At `4x` and above, however, the bonus
will decrease by `4`.  For example, `4x` will be reduced to `0x`.

The aim is to further decrease the punishment felt for getting a high
bonus.  Effectively, this gives the player the opportunity to recoup
some gold; assuming, of course, that they can afford to continue to use
Extra Chance.