[Src/Error] Add `Error` and `Label`
In the vein of `errno`, there are now `Error` and `Label` globals.

The purpose of `Error` is to specify `string` error messages.  The
purpose of `Label` is a bit more nuanced.  It goes hand in hand with
`Error`, but is intended to be used to label the currently executing
thread.