[Src] Add Trigger library
The primary point of this library is to introduce a try/catch mechanism
for `trigger`.  This is achieved by using `Or` in conjunction with
`TriggerEvaluate`.  The specified try function is bound to a catch
function that is only evaluated should the thread crash or `false` be
returned.  If no error is encountered, performance is no worse than a `
trigger` using `TriggerEvaluate` without the catch mechanism.  Note that
this requires the try function to explicitly return `true` upon success.
Use of the library is required to destroy or clear a `trigger` utilizing
try/catch, else the paired try/catch `boolexpr` will leak.