[Src/Run] Add scheduler and thread crash detection
This commit deals primarily with making it easier to detect errors,
specifically thread crashes.  In regards to the scheduler, it leverages
a high frequency timer.  This seems to exhibit superior accuracy to a
long duration timer and getting its elapsed time.

Now, performance is not on par with native timers.  But, it is difficult
to maintain a feature set that encompasses the above stated accuracy,
the ability to detect thread crashes, and the exposing of an index to
easily identify individual scheduled functions.