[Gem] Add Immolation library
Issue: #55

This provides a quick replacement for Immolation, albeit with a big
limitation: only one Immolation damage source can exist per unit.  This
is fine for Gem.

It works by having each damage source apply damage.  That is the easy
part.  The more annoying part is that too many effects on the screen
will tank performance.  So we register each damage source to an effect
group.  A single effect per period (we assume that each group's damage
sources will all share the same period, but they may not) per group.
Effectively, this is almost visually similar to the bugged version of
Immolation; except, of course, that the it is group wide instead of
across all Immolation sources.