[Src] Add and use Zeta library
Yeah, it is a dumb name.  This entire topic of conversation shouldn't be
necessary, but it is.  In short, 'fake' official versions hurt the
community.  In reality, there is not much that can be done other than
informing people why the Gem community would prefer people avoid
releasing fake versions.  And so here we are.