[Gem] Rework swap
Issue: #113

For starters, the ability to swap will no longer be on individual
towers.  Rather, it will be a core ability of the builder.  It will also
no longer cost gold.  Instead, it will cost mana, which is only
recovered after one finishes a round.

It also functions differently.  The first target of the pair to be
swapped is marked.  A subsequent usage of swap will then swap the second
target with the first.