Elder: Fix spells not casting
Issue: #441

This fixes an issue in 4c7d8043ba7e4be592195956be0a9e36b29e65ce.

Basically, I did not test that Elder Slate actually worked.  I was lazy,
and tried to use the existing code.  That's the first problem.

The second is that using the on damage event in conjunction with Elder
Slate will not work in patch 1.30.4.  Elder uses Sapphire damage, which
is just spell damage in attack form.  However, damage detection cannot
tell the difference between spell damage from an attack and that from an
ability.  As such, we have to revert to the on attack event until patch
1.31, which provides additional damage event functionality.