Rank: Update comparison rules
The existing rules first compare level progression, then damage done on
the furthest round reached, and then, in rare event of a tie, the finish
time for the previous round.  The full check mostly applies to the
Damage Test, which is internally considered Level 51, as players are on
Level 52 upon killing it and are guaranteed to have done the same
damage.  As damage is stored as a `real`, the case of two players having
done equal damage when leaking is exceedingly rare.

Still, in a push to allow easier comparison across multiple games, as
well as improving handling in the rare situations where ties can occur
in-game, this is being adjusted.  If damage done is equal, then the stop
time (which can include both the finish time as well as the time at
which a player lost) is checked for all rounds, starting from the most
recent.  If a tie is still found, then two players are equal in terms of
the comparison.