Board: Display damage done upon losing
Issue: #439

Upon losing, a player's damage done during a round will be shown to them
in the title.  It will replace Extra Chance information, which mostly
serves no purpose once a player is unable to continue.

The point behind this change is to allow comparison of players who have
leaked on the same level across multiple games.