Gem: Remove lives
Issue: #435

This removes the old lives mechanic in its entirety.  It wasn't very
intuitive for new players, and being able to buy new lives did not sit
well with a lot of good players.  So, it is replaced.

From now on, any leak on Level 10 and beyond will result in a loss.
Leaks before Level 10 are not penalized, beyond the increase in time and
loss of bounty gold that they represent.  These are pretty simply rules
to understand.

As far as why no penalty for leaking before Level 10, it is simple.
Leave the early game noob friendly.  They now have more time to mess
around and figure things out, before being summarily shown the door.
And, it does not represent a balance issue.  Yes, people could leak far
more than they did before, recover, and finish the game.  But they will
not be setting any records.