Gem: Indicate when a player has lost
Issue: #422

This issue was tackled previously by turning lives to red.  But, with
the removal of lives that is not longer an option.  Instead, we opt to
no longer indicate that a player has lost on the board at all.  An
astute player that cares if someone else has lost would either notice
the message indicator such event has happened or realize that a gold
mine is missing.  For the player that actually lost, their builder is
renamed to `Game Over`.