Gem: Push Damage Test into 50
Issue: #90

This has been a long time coming.  To sum up the major issue, the
stunlock mechanic that Ancients provide effectively eliminates using a
single unit as any sort of meaningful measuring tool.  Additionally, I
would rather not, at this juncture, change the mechanic entirely.

With that out of the way, the Damage Test (i.e. the Boss) will spawn
alongside his minions on Level 50.  Obviously, the Boss has to have
reduced hit points.  For the dominant Ancient based strategy, this will
actually be more difficult to complete than a single unit with an absurd
amount of hit points.

This change is intentionally an indirect nerf to Ancients, as it will
overwhelm them within Level 50 itself while the minions are still alive.
This increases the liklihood of leaking, especially on shorter mazes,
and thus makes the final challenge of the game rather difficult.  On top
of that, it opens up other non-Ancient focused strategies, as anything
else was incapable of killing the two million hit points boss.