Dark Emerald: Do not increase stun chance with count
Due to bounce not coming back to the same targets, the idea of having an
increased stun chance on a single target was compensated for by making
the base stun chance grow.  However, this was found to be too effective
at stunning groups.  So, we revert back to the old `12.5%` chance to
stun.