Spawn: Check creep HP on map start
Issue: #448

This fixes a desync added in ae699900e66695af9e10b2bfa329cbcbbcc97937.

In short, the added function to get the total HP of a round was creating
units.  This function was being called from within a `GetLocalPlayer
()`.  This was the desync.

In another matter, if a thread crashes within a `GetLocalPlayer ()`
block, it must crash for all players or there will be a desync.  Don't
ask how I know.  I just know.