Ancient: Rework stun duration
Issue: #253

The timings on Ancients have changed.  Previously, there were `2.5`
seconds for the stun and `5.0` seconds for the disarm.  This meant that
the value of a single Ancient, in regards to stun, was rather low.  A
second or third Ancient was needed, in a cluster, to truly realize the
value of the stun they provide.

This commit reduces the disarm time to `2.5` seconds, and also changes
the stun duration to be variable, from `1.5` seconds to `3.5` seconds.
This chance increases the likelihood that a single Ancient will be able
to have chain stuns.  Thus, a single Ancient has more value than before.
A cluster still increases the odds of having a longer chain of stuns.
However, the option to not cluster them is now on the table.