Builder: Add inventory and new player tips
Issue: #93
Issue: #102

The idea is to give new players a bit of a push in the right direction,
answering the most basic and common questions frequently encountered.
This is not a tutorial.  But, the hope is that nine levels where leaks
do not matter, watching other players, and messing around will encourage
new players to stick around, at least for a few more games.