Dark Emerald: Enable bounce on attack by default
This adds a bounce to all Dark and Enchanted Emerald attacks.  This
bounce will continue to grow as more towers are added.  The idea is to
increase the value of Dark Emerald.

Now, there is a peculiarity with the implementation used.  When a
critical strike occurs there is no bounce.  This is how Warcraft III
critical strike and bounc attack interact together.  A custom system
would be needed to avoid this.  We go with simple.