Dark Emerald: Add bonus damage while stunned
If the stun source is a Dark Emerald then the target will take bonus
damage while stunned.  This is a damage multiplier.  The idea is that
this cannot stack with the armor reduction of Ancient Slate, and thus
presents a competitive option that can help scale spell damage as well
as physical damage.