Gem: Remove collision from the monsters
Issue: #7
Issue: #244
Issue: #373
Issue: #399

Okay, so this is a simple change.  But it has a fairly large impact.
For starters, it simplifies a number of topics in the above issues.  The
monsters should no longer get confused and begin to attack towers
(except when a player has truly blocked).  No longer will they get stuck
on each other.  This allows remove of the anti-stuck ability.

Now, this could present a few changes to balance.  Slows and stuns, in
particular, may be slightly less impactful as the monsters will no
longer bunch up.  This might make some AoE damage sources a bit less
valuable.  But that's okay.