Spawn: Add additional boss units
Issue: #394

This introduces three categories of boss units, all of which are
cosmetic: normal, champion, and recognized.  Most of the time a boss
will be one of the old bosses (i.e. Damage Tests or the ReRunner).
Sometimes it will be a Gem TD+ Champion.  And rarely it will be someone
recognized for their contribution to the map.

Do note that the label/name needs to have a dash added in order to have
some separation from the level.  It just looks better.