Gem 3.1: Ignore units passing by the mine
Issue: #458

This is similar to an issue that dealt with units passiny back through
the spawn location.  Their pathing would then break.  However, by
passing too close to the mine they are simply handled as a leak, even if
their pathing status did not involve them being sent to the mine.