Uranium: Fix the slow aura value
It seems that this has been set to `75%` since Gem 3.1.  I guess there
is one huge benefit from doing the tooltips.  I essentially forced
myself to audit most of the objects.  Yeah, I probably should have done
that before now.  But hey, if it ain't broke (that you're aware of)...