Board: Improve board information for replays
Issue: #466

Display the damage done after the progress percentage.  This only
appears in replays; but, it is invaluable.  My brief testing leads me to
believe that it is not possible to utilize the multiboard title to
display data from the perspective selected when watching.  As such, it
has to be on the board proper.