Commands: Add `-bars` command to toggle status bars
Issue: #469

This introduces a new command to toggle status bars on player owned
towers.  This extends to everything but the builder, the gold mine, and
the recipe buildings. The aim is to allow the player to view the status
bars of the monsters without the clutter of their own towers.

By default, status bars on player owned towers are now hidden.