Board: Ensure progress never reaches `100%`
Issue: #487

A progress of `100%` on the board implies that the round was completed.
However, it can reach this value even if monsters are alive or leak, due
to rounding.  We ensure that `100%` simply shows `99.9%`, as when a
round is truly completed the player advances to the next one and `0%` is
shown.