Swap: Remove delay after first phase
Issue: #500

Both phases of swap used the `EFFECT` event.  However, this event in
conjunction with the abilities based on Channel caused a mandatory
delay.  Perhaps switching to a different base ability would have
resolved the issue.  Instead, use of the `CAST` event on the first phase
of the move was used.  This allows nearly instant swaps, assuming the
player is quick enough.