Gem: Remove use of `udg_CreepOwner`
Issue: #511

The belief is that there is some sort of change to event behavior in
1.31.  Previous fixes support this notion, as various events that
previously worked fail in the latest patch under various conditions.
However, this won't be tested for this particular case.  And this hunch
could be off base.

Regardless, the observed indicates that a unit can die before the event
indicating that it has entered a region fires.  Gem relied upon this
entering event for bookkeeping, and subsequently would fail if a unit
died too quickly.

The fix removes reliance on the `udg_CreepOwner` array, and instead uses
the fairly new guarantee of a one-to-one relationship between human and
computer players to determine the correct human ownership of a monster.