Bars: Fix chain upgrading
Issue: #515

If a player had enough gold, a tower would upgrade multiple times.  This
only applied to Silver, Malachite, and Star Ruby.  This bug was
introduced in 4ce87f14ba3bd5443e84c88de5b12cd020e339b5.  When behavior
was split out of the file an associated event was not disabled.