Run: Switch from a timed trigger event to timer
Issue: #513

A bug related to timed trigger events and the FPS slider requires
switching to timers, which do not express the issue.  This is not ideal,
as we lose the trigger try/catch mechanism around the `Run___Tick`
function.  However, there have been no errors reported in all of 1.6, so
we can assume that it is fine to go without that safety in place.