Swap: Fix swap having no cost at round finish
A previous change to swap switched the A part from `EFFECT` to `CAST`.
However, the change of ownership at the start of the build phase was not
taken into consideration, and as such the A part can technically fail
to reach `EFFECT` due to the target being invalid.  This results in no
mana being deducted.

The fix is to simply deduct mana at `CAST`.  Also, incrementing mana is
switched to after the `ForGroup` used to change ownership, to ensure
that it happens after the `CAST` event of A part.