Revert "[Gem] Display error on reaching `texttag` limit"
Issue: #526

This reverts commit 79350ee7dbd861338cd7c13b566060476f91049e.

As of 1.32, these checks are pointless.  In fact, they were probably
never needed.  They were added to help determine if there was something
that could be done regarding `textag` that did not appear properly.
Turns out that Blizzard needed to, and did, resolve that issue.

However, these checks do run when a player was using Alt-Tab in 1.32.
My belief is that the game will generate empty or useless `texttag` with
an id of `0` in such a state to save on performance.